requirements.txt
Package         Version
--------------- -----------
fredapi         0.5.2
joblib          1.5.2
numpy           2.3.5
pandas          2.3.3
pip             24.0
python-dateutil 2.9.0.post0
python-dotenv   1.2.1
pytz            2025.2
scikit-learn    1.7.2
scipy           1.16.3
six             1.17.0
threadpoolctl   3.6.0
tzdata          2025.2
